Allow empty payload in publish
With the old protobufs, the default value for not-set payload for nil, which was
an invalid one. When converting to Protobuf 3 with
02b97b8999e28aabb8b47934fa6c11071897ccd4 the default value changed from nil to
an empty string, but an empty string is effectively a valid payload (e.g. for an
unset message), so we have to remove the clause marking the value as invalid.

Fixes https://github.com/astarte-platform/astarte/issues/460

Signed-off-by: Riccardo Binetti <riccardo.binetti@ispirata.com>