Start the plugin OTP application as permanent
We want to start the plugin as permanent to make sure that if the plugin
application crashes, VerneMQ comes down with it (and gets restarted by K8s). To
do so, we need to define an erlang wrapper since right now VerneMQ assumes the
custom start function is contained in a module with the same name as the
application, which can't be an Elixir module due to naming constraints.
This should fix the situations where VerneMQ was not able to publish after a
RabbitMQ disconnection causing a plugin crash.

Signed-off-by: Riccardo Binetti <riccardo.binetti@ispirata.com>