Support device deletion
When a device is being deleted, it is also disconnected.
Disconnection due to deletion is required via RPC from
DataUpdaterPlant. After the device has been disconnected,
the broker sends a message on the internal `"/f"` (for
"farewell") topic to signal that the emission of messages
from the device has ended. Then, the deletion ack for
the device is written to database.
A device under deletion cannot be allowed to reconnect
until deletion is completed. This is made sure by
checking the database in the `auth_on_register` plugin
hook.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>